+++ src/domain/movies.ts
@@ -18,6 +18,16 @@ const BASE_URL = 'https://api.themoviedb.org/3';
 const API_KEY = process.env.API_KEY;
 const store: Store = Store.getInstance();
 
+export const searchMovieByKeyword = (keyword: string) => {
+
+  if (keyword === "") return;
+
+  store.resetMoviesAndPages();
+  store.setLastKeyword(keyword);
+
+  updateMovies(keyword);
+}
+
 export const checkIntersectionObserverEntries = (intersectionObserverEntries: IntersectionObserverEntry[]) => {
   const isElementVisible = intersectionObserverEntries[0].isIntersecting; // 옵저빙을 하나만 했으므로 굳이 배열을 돌리지 않습니다.
   if (isElementVisible) {
